WIP, currently doesn't protect against infloop
A "%" in an Org heading with org-sticky-header-full-path set to nil
causes an infinite loop.  This is partially a bug in Emacs (see
<https://debbugs.gnu.org/cgi/bugreport.cgi?bug=26586#13>), but the "%"
is a special character in header lines and needs to be escaped anyway.
This commit tries to do that while preserving the text properties of the
"%" from "(org-get-heading t t)", but somehow an infinite loop happens
anyway.  I'm guessing there are more "%" in the text properties (which
seem extensive for links), and I guess that's causing the problem.

I'll leave this branch for future work, but for now I'm going to simply
convert the links to plain text, which is what org-get-outline-path does
anyway.